Day9/video-avatars-agent/mcp/requirements.txt
google-adk==1.17.*
google-cloud-aiplatform==1.125.*
google-cloud-storage
google-genai==1.48.*
httpx==0.28.*
fastmcp==2.13.*
